refactor(frontend): split address services into networks
