+++ src/frontend/src/lib/components/guard/AddressGuard.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
-	import { validateBtcAddressMainnet, validateEthAddress } from '$lib/services/address.services';
 	import { initSignerAllowance } from '$lib/services/loader.services';
 	import { btcAddressMainnetStore, ethAddressStore } from '$lib/stores/address.store';
+	import { validateEthAddress } from '$eth/services/eth-address.services';
+	import { validateBtcAddressMainnet } from '$btc/services/btc-address.services';
 
 	let signerAllowanceLoaded = false;
